Updated name of Oregontrail1 to OregonTrain1
We are supposed to use Camelcase, so I updated the filename.
