+++ oregontrail1/nbproject/project.properties
@@ -54,7 +54,7 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
-main.class=oregontrail1.Oregontrail1
+main.class=oregontrail1.OregonTrail1
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
